chore(release): finalize production readiness and fix build conflicts
- docs: redesign README.md with new layout, tech stack, and APK download link
- fix(dart): resolve double cast error in order_screen and int parse in resilient_call
- fix(android): resolve Kotlin metadata and JVM target conflicts between sentry_flutter and mobile_scanner plugins
- chore(android): configure Kotlin plugin version 2.2.20 and force languageVersion 1.9 for subprojects
- chore(vscode): add launch.json to automatically inject .env variables during debug
- build: setup key.properties template for release APK signing

+++ wakeupsocial_apk/lib/features/order/screens/order_screen.dart
@@ -344,7 +344,7 @@ class _OrderScreenState extends State<OrderScreen> {
                     ),
                   ),
                   Text(
-                    _formatPrice(item.price * item.quantity),
+                    _formatPrice((item.price * item.quantity).toDouble()),
                     style: const TextStyle(
                       fontSize: 13,
                       fontWeight: FontWeight.w500,
